+++ store/app/src/main/java/br/com/stone/vianna/starstore/di/Modules.kt
@@ -3,16 +3,16 @@ package  br.com.stone.vianna.starstore.di
 import android.arch.persistence.room.Room
 import br.com.stone.vianna.starstore.AppDatabase
 import br.com.stone.vianna.starstore.BuildConfig
-import br.com.stone.vianna.starstore.view.card.*
-import br.com.stone.vianna.starstore.view.itemList.*
-import br.com.stone.vianna.starstore.view.shoppingCart.ShoppingCartContract
-import br.com.stone.vianna.starstore.view.shoppingCart.ShoppingCartPresenter
-import br.com.stone.vianna.starstore.view.shoppingCart.ShoppingCartRepository
-import br.com.stone.vianna.starstore.view.shoppingCart.ShoppingCartRepositoryImpl
-import br.com.stone.vianna.starstore.view.transactionHistory.TransactionContract
-import br.com.stone.vianna.starstore.view.transactionHistory.TransactionPresenter
-import br.com.stone.vianna.starstore.view.transactionHistory.TransactionRepository
-import br.com.stone.vianna.starstore.view.transactionHistory.TransactionRepositoryImpl
+import br.com.stone.vianna.starstore.feature.card.*
+import br.com.stone.vianna.starstore.feature.itemList.*
+import br.com.stone.vianna.starstore.feature.shoppingCart.ShoppingCartContract
+import br.com.stone.vianna.starstore.feature.shoppingCart.ShoppingCartPresenter
+import br.com.stone.vianna.starstore.feature.shoppingCart.ShoppingCartRepository
+import br.com.stone.vianna.starstore.feature.shoppingCart.ShoppingCartRepositoryImpl
+import br.com.stone.vianna.starstore.feature.transactionHistory.TransactionContract
+import br.com.stone.vianna.starstore.feature.transactionHistory.TransactionPresenter
+import br.com.stone.vianna.starstore.feature.transactionHistory.TransactionRepository
+import br.com.stone.vianna.starstore.feature.transactionHistory.TransactionRepositoryImpl
 import com.jakewharton.retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
