+++ store/app/src/main/java/br/com/stone/vianna/starstore/feature/card/CreditCardActivity.kt
@@ -1,4 +1,4 @@
-package br.com.stone.vianna.starstore.view.card
+package br.com.stone.vianna.starstore.feature.card
 
 import android.os.Bundle
 import br.com.stone.vianna.starstore.R
