+++ store/app/src/main/java/br/com/stone/vianna/starstore/feature/itemList/ItemListPresenter.kt
@@ -1,4 +1,4 @@
-package br.com.stone.vianna.starstore.view.itemList
+package br.com.stone.vianna.starstore.feature.itemList
 
 import br.com.stone.vianna.starstore.entity.Item
