+++ store/app/src/main/java/br/com/stone/vianna/starstore/feature/transactionHistory/TransactionAdapter.kt
@@ -1,4 +1,4 @@
-package  br.com.stone.vianna.starstore.view.transactionHistory
+package  br.com.stone.vianna.starstore.feature.transactionHistory
 
 import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
