+++ store/app/src/main/res/layout/activity_credit_card.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.card.CreditCardActivity">
+    tools:context=".feature.card.CreditCardActivity">
 
     <android.support.v7.widget.Toolbar
         android:id="@+id/cards_toolbar"
